python/roger/requirements.txt
aiogram==2.25.1
aiogram-media-group==0.5.1
aiohttp==3.8.3
aiosignal==1.2.0
astroid==3.0.1
async-timeout==4.0.2
asyncio==3.4.3
attrs==22.1.0
autopep8==2.0.4
Babel==2.9.1
beautifulsoup4==4.11.1
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
DateTime==4.7
decorator==5.1.1
dill==0.3.7
dnspython==2.2.1
frozenlist==1.3.1
future==0.18.2
geocoder==1.38.1
idna==3.4
isort==5.12.0
linkpreview==0.2.2
magic-filter==1.0.12
mccabe==0.7.0
multidict==6.0.2
openai==0.27.0
platformdirs==4.0.0
pycodestyle==2.11.1
pylint==3.0.2
pymongo==4.3.2
pyTelegramBotAPI==4.7.0
python-dateutil==2.8.2
pytz==2022.5
ratelim==0.1.6
requests==2.28.1
schedule==1.1.0
six==1.16.0
soupsieve==2.3.2.post1
tomli==2.0.1
tomlkit==0.12.2
tqdm==4.64.1
typing_extensions==4.8.0
urllib3==1.26.12
yarl==1.8.1
zope.interface==5.5.0
